revert changes to the TIDARoiDescriptor - CTP size of L1ID is only 32b
Former-commit-id: 6be88822184cb3301974165b51967e95d98f5176

+++ Trigger/TrigAnalysis/TrigInDetAnalysis/src/TIDARoiDescriptor.cxx
@@ -43,7 +43,7 @@ TIDARoiDescriptor::TIDARoiDescriptor( double eta,  double etaMinus,   double eta
 { }
 
 
-TIDARoiDescriptor::TIDARoiDescriptor( unsigned long long l1id, int id, 
+TIDARoiDescriptor::TIDARoiDescriptor( unsigned int l1id, int id, 
 				      double eta,  double etaMinus,   double etaPlus,
 				      double phi,  double phiMinus,   double phiPlus,
 				      double zed,  double zedMinus,   double zedPlus )
@@ -55,8 +55,7 @@ TIDARoiDescriptor::TIDARoiDescriptor( unsigned long long l1id, int id,
 { }
 
 
-TIDARoiDescriptor::TIDARoiDescriptor(unsigned int roiword, unsigned long long l1id,
-				     int id, 
+TIDARoiDescriptor::TIDARoiDescriptor(unsigned int roiword, unsigned int l1id, int id, 
 				     double eta,  double etaMinus,   double etaPlus,
 				     double phi,  double phiMinus,   double phiPlus,
 				     double zed,  double zedMinus,   double zedPlus )
